+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * (Early Access) Configurable internal cluster security allows users to configure encryption and authentication on the internal connections within the Apache Kafka cluster.
 * Update Kafka Exporter to [1.10.0](https://github.com/danielqsj/kafka_exporter/releases/tag/v1.10.0)
 * Support for mounting projected service account tokens into Strimzi-managed Pods
+* Prevent removal of JBOD volumes which still contain partition replicas. The check can be disabled with the `strimzi.io/skip-broker-scaledown-check` annotation on the `Kafka` resource.
 
 ### Major changes, deprecations, and removals